+++ vm/src/arch/integration_api.rs
@@ -26,7 +26,9 @@ pub trait MachineAdapterInterface<T> {
 /// The adapter owns all memory accesses and timestamp changes.
 /// The adapter AIR should also own `ExecutionBridge` and `MemoryBridge`.
 pub trait MachineAdapter<F: PrimeField32> {
+    /// Records generated by adapter before main instruction execution
     type ReadRecord;
+    /// Records generated by adapter after main instruction execution
     type WriteRecord;
     type Air: BaseAir<F>;
     type Cols<T>;
